+++ ChangeLog.md
@@ -2,6 +2,10 @@
 
 ## Unreleased changes
 
+## 0.1.9
+
+* 
+
 ## 0.1.8
 
 * Changed `insertIfExists` to `add`.